Move all android plugin code to new folder structure
