+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'camera'
+rootProject.name = 'flutter_better_camera'
