+++ android/src/main/java/app/lightsnap/flutter_better_camera/CameraPermissions.java
@@ -1,11 +1,13 @@
-package io.flutter.plugins.camera;
+package app.lightsnap.flutter_better_camera;
 
 import android.Manifest;
 import android.Manifest.permission;
 import android.app.Activity;
 import android.content.pm.PackageManager;
+
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
+
 import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugin.common.PluginRegistry.RequestPermissionsResultListener;
