+++ android/src/main/java/app/lightsnap/flutter_better_camera/DartMessenger.java
@@ -1,4 +1,4 @@
-package io.flutter.plugins.camera;
+package app.lightsnap.flutter_better_camera;
 
 import android.text.TextUtils;
 import androidx.annotation.Nullable;
