+++ example/android/app/build.gradle
@@ -32,7 +32,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "io.flutter.plugins.cameraexample"
+        applicationId "app.lightsnap.flutter_better_camera_example"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
